backend/requirements.txt
asgiref==3.6.0
Django==4.1.4
sqlparse==0.4.3

Pillow==9.3.0

djangorestframework==3.14.0
pytz==2022.7
Markdown==3.4.1
django-filter==22.1

djoser==2.1.0

# удалить
# djangorestframework-simplejwt==5.2.2

drf-extra-fields==3.4.1
